requirements.txt
fastapi
uvicorn
websockets
websocket-client
opencv-python
numpy
torch
ultralytics
pydantic